+++ create_backend_executable.sh
@@ -1,5 +1,5 @@
 #! /bin/bash
 cd SomnusBackend
 sudo mvn clean -Pproduction package
-cp target/somnus_server-v*.jar deploy
+cp target/somnus_server-v*.jar .
 cd ..