+++ SomnusFrontEnd/src/app/components/server-stats/server-stats.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { AuthenticationService } from 'src/app/services/authentication/authentication.service';
 
 @Component({
   selector: 'app-server-stats',
@@ -10,9 +11,15 @@ export class ServerStatsComponent implements OnInit {
   // app user
   user: firebase.default.User;
 
-  constructor() { }
+  constructor(
+    private authenticationService: AuthenticationService,
+  ) { }
 
   ngOnInit(): void {
+    this.authenticationService.getLoggedInUser()
+      .subscribe (user => {
+        this.user = user;
+    });
   }
 
 }