+++ changelog.md
@@ -1,7 +1,15 @@
 # Changelog
 All changes made to this project will be documented in this file
 
-## [Unreleased]
+## [RELEASED]
+---------------------------------------------------------
+## [0.2.1] - 2021-02-11
+### Added
+- Properties for security
+- Release info
+
+### Changed
+- Minor changes in Contributions Service
 ---------------------------------------------------------
 ## [0.2.1] - 2021-02-11
 ### Added